{"schema":"decodex/commit/1","summary":"Extract frozen overlay export bridge owner","authority":"manual"}
